+++ app/src/main/java/com/baec23/upcycler/model/ChatSession.kt
@@ -1,14 +1,14 @@
 package com.baec23.upcycler.model
 
 data class ChatSession(
-    val chatSessionId: Int = 0,
-    val jobCreatorUserId: Int = 0,
+    val chatSessionId: Long = 0,
+    val jobCreatorUserId: Long = 0,
     val jobCreatorDisplayName: String = "",
-    val workerUserId: Int = 0,
+    val workerUserId: Long = 0,
     val workerDisplayName: String = "",
-    val jobId: Int = 0,
+    val jobId: Long = 0,
     val jobImageUrl: String = "",
     var mostRecentMessage: String = "",
     val mostRecentMessageTimestamp: Long = 0,
-    val participantUserIds: List<Int> = emptyList()
+    val participantUserIds: List<Long> = emptyList()
 )
