+++ app/src/main/java/com/baec23/upcycler/model/Job.kt
@@ -1,8 +1,8 @@
 package com.baec23.upcycler.model
 
 data class Job(
-    val jobId: Int = 0,
-    val creatorId: Int = 0,
+    val jobId: Long = 0,
+    val creatorId: Long = 0,
     val createdTimestamp: Long = 0L,
     val title: String = "",
     val details: String = "",
